@@ -1,31 +1,33 @@
 {
-  "name": "vite-project",
-  "version": "0.0.1",
+  "name": "vite",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "start": "vite",
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src",
-    "format": "prettier ./src  --write"
+    "format": "prettier .  --write"
   },
   "dependencies": {
-    "axios": "^0.24.0",
-    "vue": "^3.2.25",
-    "vue-axios": "^3.4.0",
-    "vue-router": "^4.0.12",
-    "vuex": "^4.0.2"
+    "axios": "1.1.2",
+    "install": "0.13.0",
+    "npm": "8.19.2",
+    "pinia": "2.0.23",
+    "vue": "3.2.37",
+    "vue-router": "4.1.5"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^2.0.0",
-    "autoprefixer": "^10.4.0",
-    "eslint": "^8.16.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^9.0.1",
-    "postcss": "^8.4.5",
-    "prettier": "^2.6.2",
-    "tailwindcss": "^3.0.1",
-    "vite": "^2.7.1"
+    "@vitejs/plugin-vue": "3.1.0",
+    "autoprefixer": "10.4.12",
+    "eslint": "8.25.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-vue": "9.6.0",
+    "postcss": "8.4.18",
+    "prettier": "2.7.1",
+    "tailwindcss": "3.1.8",
+    "vite": "3.1.0"
   }
 }
