@@ -1,23 +1,13 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-const path = require('path')
+import path from 'path'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [vue()],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, 'src')
+      '@': path.resolve(__dirname, 'src'),
     },
   },
-  productionSourceMap: process.env.NODE_ENV != 'production',
-  chainWebpack: config => {
-    config.optimization.minimizer('terser').tap((args) => {
-      args[0].terserOptions.output = {
-        ...args[0].terserOptions.output,
-        comments: false
-      }
-      return args
-    })
-  }
 })
