@@ -2,9 +2,13 @@ module.exports = {
   env: {
     node: true,
   },
-  extends: ['eslint:recommended', 'plugin:vue/vue3-recommended', 'prettier'],
+  extends: [
+    'eslint:recommended',
+    'plugin:vue/vue3-recommended',
+    'prettier'
+  ],
   rules: {
     // override/add rules settings here, such as:
     // 'vue/no-unused-vars': 'error'
-  },
+  }
 }
