@@ -3,21 +3,12 @@
     <div>
       <img alt="Vue logo" src="@/assets/logo.png" />
       <HelloWorld msg="Hello Vue 3 + Vite" />
-      <router-view></router-view>
     </div>
   </div>
 </template>
 
-<script>
+<script setup>
 import HelloWorld from '@/components/HelloWorld.vue'
-
-export default {
-  name: 'LayoutMain',
-
-  components: {
-    HelloWorld,
-  },
-}
 </script>
 
 <style>
