@@ -5,6 +5,4 @@
   </div>
 </template>
 
-<script>
-export default {}
-</script>
+<script setup></script>
